Move audio resources into the .pak file.
This is needed to keep part 2's APK size below 50MB (it has
a lot of audio).

To make this work, we use a new Cricket Audio API which allows
us to use file offsets. For *that* to work, we refactor the .pak
reader to expose the offsets, and move it from APKit into glue
so that glue/AVFoundation can depend on it. Whew!