Don't align structs, to fix x86 crash
I'll write a proper commit message, as this is a sneaky one...

RATIONALE

Whenever I tried out an x86 build, it would reliably SEGV early
on. Performance with the ARM build seemed okay on x86 devices
so I decided to ignore x86 for the time being.

However, all of the people complaining of sparodic crashes in
80 Days seem to be using x86 tablets. I can only guess this is
because the system's dynamic ARM-to-x86 translator is flaky.
So a proper x86 build seems like a good way to fix this.

THE BUG

The seg faults always occurred in methods returning a struct.
In these cases, the caller reserves memory for the return value
and passes in a pointer for the callee to fill out. It turns out
that in x86, the callee was using a MOVAPS instruction which
requires 16-byte alignment. Evidently the caller wasn't always
aligning correctly, resulting in a SEGV (confusingly reported
as address 0, misleading me into thinking it was a NULL pointer).

Someone else hit a similar bug: http://www.bailopan.net/blog/?p=7

In fact the structs *were* declared with 16-byte alignment
(syntax is "_attribute__((aligned(16)))"), so that must be why
the callee was using MOVAPS. The caller failing to align them
properly must be a compiler bug, unless I'm missing something.

THE FIX

The fix is just to remove all the alignment attributes. This
will probably generate slightly slower code, but at least it
runs.

FURTHER NOTES

I might just be missing some important compiler flag, but it's
impossible to figure out because the Clang docs are fucking
useless. The nearest thing to a list of all flags is just to
run "clang -cc1 -help", as per this Stack Overflow answer:

http://stackoverflow.com/questions/7880812/complete-list-of-clang-flags

I mentioned the NULL pointer red herring above. There's a
related bug wherein calls to 'nil' that return a struct don't
properly zero out all the struct fields. I thought perhaps
there might be some general problem with struct-calls-to-nil
that was causing the crash on x86, but that doesn't seem to be
the case.