Make AP_UserDefaults more robust
AP_UserDefaults was updating its file with atomically:YES, but this
isn't actually atomic unless we also fsync() the parent directory.
And there's no guarantee that fsync() actually does anything.

Therefore, try a totally new approach: write a new file each time
the defaults change, with an incrementing sequence number so we
know which is the latest file.

To save space, we gzip the files, and erase the oldest file after
writing more than 10 files. (Possible risk: could the file erasure
be flushed before any of those new files are written? Surely not...)