Possible fix for bizarre 'diagonals' issue? Don't repeat indices in GL_TRIANGLES arrays
In a few places in the code, I draw quads as two triangles with vertices ABC-CDA.
My theory is that some broken OpenGL drivers are incorrectly interpreting the C-C
sequence as a degenerate triangle, so the second triangle is skipped.

This commit changes the sequence to ABC-DAC, which has the same effect but avoids
the stuttering.